% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_functions.R
\name{bids_camelcase}
\alias{bids_camelcase}
\title{Convert a string to BIDS-compatible camelCase}
\usage{
bids_camelcase(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A character string in camelCase form.
}
\description{
Treats periods, hyphens, and underscores as word boundaries, removes them,
and capitalizes the following letter. E.g., "task-ridl_name" becomes
"taskRidlName", and "cor.shrink" becomes "corShrink".
}
\examples{
\dontrun{
  bids_camelcase("task-ridl_name")
  bids_camelcase("echo_time-series")
  bids_camelcase("cor.shrink")
  bids_camelcase("space-mni152nlin2009casym")
}

}
\keyword{internal}
