% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pipeline.R
\name{get_all_nodes}
\alias{get_all_nodes}
\title{helper for recursively collecting all nodes from a tree}
\usage{
get_all_nodes(node)
}
\arguments{
\item{node}{A data.tree Node object}
}
\value{
A list of all descendant nodes including the node itself
}
\description{
helper for recursively collecting all nodes from a tree
}
\keyword{internal}
