% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pipeline.R
\name{get_job_type}
\alias{get_job_type}
\title{helper for getting job type from job name}
\usage{
get_job_type(job_name)
}
\arguments{
\item{job_name}{Character string job name}
}
\value{
Character string job type
}
\description{
helper for getting job type from job name
}
\keyword{internal}
