% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pipeline.R
\name{get_status_color}
\alias{get_status_color}
\title{helper for coloring status text based on status}
\usage{
get_status_color(status)
}
\arguments{
\item{status}{Character string job status}
}
\description{
helper for coloring status text based on status
}
\keyword{internal}
