% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_functions.R
\name{get_step_title}
\alias{get_step_title}
\title{helper for retrieving a human-readable title for a pipeline step}
\usage{
get_step_title(node)
}
\arguments{
\item{node}{A data.tree Node object representing a job}
}
\value{
A string summarizing the job
}
\description{
helper for retrieving a human-readable title for a pipeline step
}
