% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_tracking_functions.R
\name{get_tracked_job_status}
\alias{get_tracked_job_status}
\title{Query job status in tracking SQLite database}
\usage{
get_tracked_job_status(
  job_id = NULL,
  return_children = FALSE,
  return_parent = FALSE,
  sequence_id = NULL,
  sqlite_db
)
}
\arguments{
\item{job_id}{The job id for which to retreive the status}

\item{return_children}{Return child jobs of this job}

\item{return_parent}{Return parent jobs of this job}

\item{sequence_id}{The sequence id to query instead of \code{job_id}}

\item{sqlite_db}{Character string of sqlite database}
}
\value{
An R data.frame version of the tracking database
}
\description{
Query job status in tracking SQLite database
}
\details{
Status reads wait up to 10 seconds for a concurrent SQLite writer
before reporting a database lock error.
}
