% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_functions.R
\name{insert_df_sqlite}
\alias{insert_df_sqlite}
\title{helper function to insert a keyed data.frame into the sqlite storage database}
\usage{
insert_df_sqlite(
  gpa = NULL,
  id = NULL,
  session = NULL,
  run_number = NULL,
  data = NULL,
  table = NULL,
  delete_extant = TRUE,
  append = TRUE,
  overwrite = FALSE,
  immediate = FALSE
)
}
\arguments{
\item{gpa}{A \code{glm_pipeline_arguments} object used to lookup location of SQLite database for this insert}

\item{id}{the id of the subject to whom these data belong}

\item{session}{the session of these data}

\item{run_number}{the run_number of these data}

\item{data}{A \code{data.frame} containing the data to be inserted into the sqlite db}

\item{table}{A character string of the table name to be modified}

\item{delete_extant}{Whether to delete any existing records for this id + session + run_number combination}

\item{append}{Whether to append records to the table (passed through to dbWriteTable)}

\item{overwrite}{Whether to overwrite the existing table (passed through to dbWriteTable)}

\item{immediate}{Whether to open unique connection, commit transaction, then close the connection.
This should be useful for SQLite concurrency issues in a parallel compute environment, but at present
we are still getting errors even with the immediate approach.}
}
\value{
a TRUE/FALSE indicating whether the record was successfully inserted
}
\description{
helper function to insert a keyed data.frame into the sqlite storage database
}
\keyword{internal}
