% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_4d_nifti.R
\name{interp_4d_nifti}
\alias{interp_4d_nifti}
\title{Interpolate a 4d NIFTI file}
\usage{
interp_4d_nifti(
  in_file = NULL,
  out_file = NULL,
  timepoints = NULL,
  edge_nn = TRUE
)
}
\description{
Interpolate a 4d NIFTI file
}
