% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{is_external_path}
\alias{is_external_path}
\title{Determine if a path lies outside the project directory}
\usage{
is_external_path(path, project_dir)
}
\arguments{
\item{path}{The path to evaluate.}

\item{project_dir}{The root project directory.}
}
\value{
\code{TRUE} if \code{path} is not within \code{project_dir}, otherwise \code{FALSE}.
}
\description{
Determine if a path lies outside the project directory
}
\keyword{internal}
