% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{load_study}
\alias{load_study}
\title{Load a study configuration from a file}
\usage{
load_study(file = NULL, validate = TRUE)
}
\arguments{
\item{file}{A YAML file containing a valid study configuration.}

\item{validate}{Logical indicating whether to validate the configuration after loading. Default: TRUE}
}
\value{
A list representing the study configuration (class \code{"bg_study_cfg"}). If \code{validate} is TRUE,
the returned object is validated (missing fields may be set to NULL and noted).
}
\description{
Load a study configuration from a file
}
