% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_functions.R
\name{out_file_exists}
\alias{out_file_exists}
\title{Check for Existence of a BIDS-Formatted Output File with a given description}
\usage{
out_file_exists(in_file, description, overwrite = TRUE)
}
\arguments{
\item{in_file}{Path to the input BIDS file (e.g., a preprocessed BOLD image).}

\item{description}{Character string to use as the new \code{desc} field in the expected output file.}

\item{overwrite}{Logical. If \code{FALSE}, existing files will not be overwritten.}
}
\value{
A list with elements:
\item{out_file}{Path to the expected output file.}
\item{skip}{Logical indicating whether to skip writing due to file existence.}
}
\description{
This function constructs a BIDS-compliant filename based on an input file, replacing
the \code{desc} field with a specified \code{description}, and checks whether the corresponding output file
already exists. If the file exists and \code{overwrite = FALSE}, the function returns \code{skip = TRUE}.
}
\keyword{internal}
