% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_functions.R
\name{parse_complete_time}
\alias{parse_complete_time}
\title{Parse completion time from a .complete file}
\usage{
parse_complete_time(file)
}
\arguments{
\item{file}{Path to .complete file.}
}
\value{
POSIXct time or NA if parsing fails.
}
\description{
Parse completion time from a .complete file
}
\keyword{internal}
