% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_inventory.R
\name{print.bg_qc_inventory}
\alias{print.bg_qc_inventory}
\title{Print a compact QC snapshot summary}
\usage{
\method{print}{bg_qc_inventory}(x, ...)
}
\arguments{
\item{x}{A \code{bg_qc_inventory} object.}

\item{...}{Unused arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a compact QC snapshot summary
}
