% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pipeline.R
\name{print_step_tree_by_type}
\alias{print_step_tree_by_type}
\title{helper for printing step summaries organized by subject then job type}
\usage{
print_step_tree_by_type(tree_root)
}
\arguments{
\item{tree_root}{The root node of the sequence tree}
}
\description{
helper for printing step summaries organized by subject then job type
}
\keyword{internal}
