% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remove_nifti_volumes}
\alias{remove_nifti_volumes}
\title{Remove Specified Timepoints from a 4D NIfTI Image}
\usage{
remove_nifti_volumes(infile, remove_tpts, outfile)
}
\arguments{
\item{infile}{Character string. Path to the input 4D NIfTI file.}

\item{remove_tpts}{Integer vector. Timepoints (1-based) to remove from the image.}

\item{outfile}{Character string. Path to save the output NIfTI file with selected volumes.}
}
\value{
None. The function writes a new NIfTI file to \code{outfile}.
}
\description{
This function removes specified timepoints (volumes) from a 4D NIfTI image and saves
the resulting image to a new file. Timepoints are specified using 1-based indexing,
consistent with R conventions.
}
\details{
This function uses the \code{volumes} argument in RNifti to efficiently read
only the retained timepoints from disk. If all volumes are removed, an error is thrown.
The input image must be 4-dimensional (i.e., include a time dimension).
}
\examples{
\dontrun{
remove_nifti_volumes("input_bold.nii.gz", remove_tpts = c(1, 2, 100), 
  outfile = "trimmed_bold.nii.gz")
}

}
