% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_postprocess.R
\name{setup_apply_aroma}
\alias{setup_apply_aroma}
\title{Configure ICA-AROMA denoising application in postprocessing}
\usage{
setup_apply_aroma(scfg = list(), fields = NULL)
}
\arguments{
\item{scfg}{A study configuration object created by \code{setup_project()}.}

\item{fields}{A character vector of fields to prompt for. If \code{NULL}, all fields will be prompted.}
}
\value{
A modified version of \code{scfg} with the \verb{$postprocess$apply_aroma} entry updated.
}
\description{
This function configures the application of ICA-AROMA denoising to fMRI data as part of postprocessing.
ICA-AROMA (Pruim et al., 2015) identifies and labels motion-related independent components (ICs) using
spatiotemporal features, and outputs regressors that can be used to remove these components.
}
\details{
If enabled, this step applies the AROMA regressors to remove noise components from the BOLD time series
using either 'aggressive' or 'nonaggressive' regression. Nonaggressive denoising is recommended as it
preserves shared variance with signal components.

This step assumes that ICA-AROMA has already been run using a tool like \code{fmripost-aroma}.
}
\keyword{internal}
