% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_bids_conversion}
\alias{setup_bids_conversion}
\title{Specify the BIDS conversion settings}
\usage{
setup_bids_conversion(scfg, fields = NULL)
}
\arguments{
\item{scfg}{a project configuration object, as produced by \code{load_project} or \code{setup_project}}

\item{fields}{a character vector of fields to be prompted for. If \code{NULL}, all fields will be prompted for.}
}
\value{
a modified version of \code{scfg} with \verb{$bids_conversion} populated
}
\description{
Specify the BIDS conversion settings
}
\keyword{internal}
