% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{setup_bids_validation}
\alias{setup_bids_validation}
\title{Specify the BIDS validation settings}
\usage{
setup_bids_validation(scfg, fields = NULL)
}
\arguments{
\item{scfg}{A study configuration object, as produced by \code{load_study()} or \code{setup_study()}.}

\item{fields}{A character vector of fields to be prompted for. If \code{NULL}, all BIDS validation fields will be prompted for.}
}
\value{
A modified version of \code{scfg} with the \verb{$bids_validation} entry populated.
}
\description{
Specify the BIDS validation settings
}
\keyword{internal}
