% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_flywheel_sync}
\alias{setup_flywheel_sync}
\title{Configure flywheel sync settings}
\usage{
setup_flywheel_sync(scfg, fields = NULL)
}
\arguments{
\item{scfg}{A project configuration object, as produced by \code{load_project()} or \code{setup_project()}.}

\item{fields}{A character vector of fields to be prompted for. If \code{NULL}, all Flywheel fields will be prompted for.}
}
\value{
A modified version of \code{scfg} with the \verb{$flywheel_sync} entry populated.
}
\description{
Sets up synchronization from a Flywheel instance prior to BIDS conversion.
Prompts for the Flywheel project URL, drop-off directory for downloaded DICOMs,
and a temporary directory used during transfer. Standard job settings are also
collected through \code{setup_job}.
}
\keyword{internal}
