% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{setup_mriqc}
\alias{setup_mriqc}
\title{Specify the MRIQC settings}
\usage{
setup_mriqc(scfg, fields = NULL)
}
\arguments{
\item{scfg}{A study configuration object, as produced by \code{load_study()} or \code{setup_study()}.}

\item{fields}{A character vector of fields to be prompted for. If \code{NULL}, all MRIQC fields will be prompted for.   # \link{Added}}
}
\value{
A modified version of \code{scfg} with the \verb{$mriqc} entry populated.
}
\description{
Specify the MRIQC settings
}
\keyword{internal}
