% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_postprocess.R
\name{setup_postprocess_streams}
\alias{setup_postprocess_streams}
\title{Configure postprocessing settings for a study}
\usage{
setup_postprocess_streams(scfg = list(), fields = NULL)
}
\arguments{
\item{scfg}{A study configuration object, as produced by \code{setup_project()}.}

\item{fields}{A character vector of field names to prompt for. If \code{NULL}, all postprocessing fields will be prompted.}
}
\value{
A modified version of \code{scfg} with the \verb{$postprocess} field populated.

Modified \code{scfg} with one or more postprocessing streams
}
\description{
This function enables and configures the postprocessing steps to be applied after fMRIPrep.
Postprocessing may include denoising, smoothing, filtering, intensity normalization,
and confound regression applied to preprocessed BOLD data.
}
\details{
The function interactively prompts the user (or selectively prompts based on \code{fields})
to specify whether postprocessing should be performed, and if so, how each step should be configured.

Postprocessing is applied to the outputs of fMRIPrep to prepare BOLD time series for statistical modeling.
This may include:
\itemize{
\item Applying brain masks
\item Spatial smoothing
\item ICA-AROMA denoising
\item Temporal filtering
\item Intensity normalization
\item Confound calculation and regression
}

Each step is optional and configurable. This function sets default values for memory, runtime,
and cores, and invokes a series of sub-setup functions to collect postprocessing parameters.

Interactively manage multiple postprocessing configurations. Users can add,
edit, or delete postprocessing streams. This wrapper is called by
\code{setup_project()} and invokes \code{setup_postprocess()} for each stream.
}
\keyword{internal}
