% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_postprocess.R
\name{setup_scrubbing}
\alias{setup_scrubbing}
\title{Configure scrubbing of high-motion volumes}
\usage{
setup_scrubbing(scfg = list(), fields = NULL)
}
\arguments{
\item{scfg}{A study configuration object.}

\item{fields}{Optional vector of fields to prompt for.}
}
\value{
Modified \code{scfg} with \verb{$postprocess$scrubbing} populated.
}
\description{
Generates spike regressors based on expressions evaluated on the confounds
file (e.g., "framewise_displacement > 0.9" or "-1:1; dvars > 1.5"). These regressors can later be
used to censor volumes during modeling.
}
\keyword{internal}
