% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_postprocess.R
\name{setup_spatial_smooth}
\alias{setup_spatial_smooth}
\title{Configure spatial smoothing settings for fMRI postprocessing}
\usage{
setup_spatial_smooth(scfg = list(), fields = NULL)
}
\arguments{
\item{scfg}{A study configuration object created by \code{setup_project()}.}

\item{fields}{A character vector of field names to prompt for. If \code{NULL}, all spatial smoothing fields will be prompted.}
}
\value{
A modified version of \code{scfg} with the \verb{$postprocess$spatial_smooth} field populated.
}
\description{
This function configures the spatial smoothing step for postprocessing of BOLD fMRI data.
Spatial smoothing increases signal-to-noise ratio by averaging nearby voxels and can improve
the statistical properties of the data, especially for group-level analyses.
}
\details{
The user is asked whether they want to apply smoothing, and if so, to specify the full width at half maximum (FWHM)
of the Gaussian smoothing kernel and a filename prefix.

If enabled, spatial smoothing is applied to the preprocessed BOLD data using a Gaussian kernel
with the user-specified FWHM in millimeters. This can help improve sensitivity and inter-subject alignment,
especially in standard space. This is accomplished using FSL's contrast-sensitive susan smoothing command.
}
\keyword{internal}
