% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_functions.R
\name{sqlite_table_exists}
\alias{sqlite_table_exists}
\title{helper function to check if a table exists in a SQLite database}
\usage{
sqlite_table_exists(sqlite_db, table_name)
}
\arguments{
\item{sqlite_db}{Path to SQLite database file}

\item{table_name}{Name of the table to check for}
}
\value{
Logical indicating whether the table exists
}
\description{
helper function to check if a table exists in a SQLite database
}
\keyword{internal}
