% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_functions.R
\name{tracking_df_to_tree}
\alias{tracking_df_to_tree}
\title{helper for converting tracking data.frame into a multi-level data.tree hierarchy}
\usage{
tracking_df_to_tree(tracking_df)
}
\arguments{
\item{tracking_df}{A data.frame returned by get_tracked_job_status()}
}
\value{
A named list of data.tree Node objects (one per sequence_id)
}
\description{
helper for converting tracking data.frame into a multi-level data.tree hierarchy
}
