% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_functions.R
\name{bids_camelcase}
\alias{bids_camelcase}
\title{Convert a string to BIDS-compatible camelCase}
\usage{
bids_camelcase(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A character string in camelCase form.
}
\description{
Removes hyphens/underscores and capitalizes the letter following them.
E.g., "task-ridl_name" -> "taskRidlName".
}
\examples{
\dontrun{
  bids_camelcase("task-ridl_name")
  bids_camelcase("echo_time-series")
  bids_camelcase("space-mni152nlin2009casym")
}

}
\keyword{internal}
