% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainGnomes-package.R
\docType{package}
\name{BrainGnomes-package}
\alias{BrainGnomes}
\alias{BrainGnomes-package}
\title{BrainGnomes: An R package for preprocessing and analyzing fMRI data}
\description{
This R package handles preprocessing of fMRI data on HPC clusters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://uncdependlab.github.io/BrainGnomes/}
  \item Report bugs at \url{https://github.com/UNCDEPENdLab/BrainGnomes/issues}
}

}
\author{
\strong{Maintainer}: Michael Hallquist \email{michael.hallquist@gmail.com}

}
\keyword{internal}
