% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pipeline.R
\name{diagnose_pipeline}
\alias{diagnose_pipeline}
\title{Function for diagnosing errors in a run of the pipeline}
\usage{
diagnose_pipeline(input)
}
\arguments{
\item{input}{A character path to an scfg object or an scfg object itself.}
}
\description{
Function for diagnosing errors in a run of the pipeline
}
\author{
Zach Vig & Dan Shallal
}
