% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rois.R
\name{extract_rois}
\alias{extract_rois}
\title{Extract ROI timeseries and connectivity matrices}
\usage{
extract_rois(
  bold_file,
  atlas_files,
  out_dir,
  log_file = NULL,
  cor_method = c("pearson", "spearman", "kendall", "cor.shrink"),
  roi_reduce = c("mean", "median", "pca", "huber"),
  brain_mask = NULL,
  min_vox_per_roi = 5,
  save_ts = TRUE,
  rtoz = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{bold_file}{Path to a 4D NIfTI file containing postprocessed BOLD data.}

\item{atlas_files}{Character vector of atlas NIfTI files with integer
ROI labels.}

\item{out_dir}{Directory where output files should be written.}

\item{log_file}{If not \code{NULL}, the log file to which details should be written.}

\item{cor_method}{Correlation method(s) to use when computing functional
connectivity. Supported options include "pearson", "spearman",
"kendall", and "cor.shrink". Use "none" to skip correlation
computation. Multiple methods may be supplied.}

\item{roi_reduce}{Method used to summarize voxel time series within each
ROI. Options are "mean" (default), "median", "pca", or "huber".}

\item{brain_mask}{Optional brain mask NIfTI file. If \code{NULL}, a mask
is generated by excluding voxels with zero variance across time.}

\item{min_vox_per_roi}{The minimum number of voxels required for an ROI to
be extracted and entered into correlations. If the ROI is smaller than this,
it show up as NA in the outputs, keeping the dimensionality of the connectivity
matrix consistent across inputs. Default: \code{5}}

\item{save_ts}{If \code{TRUE}, save the ROI time series (aggregated using \code{roi_reduce} method)
to \verb{_timeseries.tsv}. files. Useful for running external analyses on the ROIs. Default: \code{TRUE}.}

\item{rtoz}{If \code{TRUE}, using Fisher's z (aka atanh) transformation on correlations to make them
continuous and unbounded, rather than \verb{[0,1]}. The diagonal of the correlation matrices beccomes
15 to approximate the 1.0 correlation, rather than making it \code{Inf}.}

\item{overwrite}{If \code{TRUE}, overwrite existing timeseries.tsv or connectivity.tsv files.}
}
\value{
A named list. Each element corresponds to an atlas and contains
paths to the written timeseries (\code{timeseries}) and correlation
matrix (\code{correlation}, or \code{NULL} if not computed).
}
\description{
Given a postprocessed BOLD NIfTI file and one or more atlas images,
this function computes the mean timeseries within each ROI and
optionally computes ROI-to-ROI correlation matrices.
}
\details{
Voxels labelled in the atlas but lying outside the brain are
automatically excluded by intersecting with a brain mask derived from
the input timeseries.
}
