% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_postprocess.R
\name{get_postproc_output_files}
\alias{get_postproc_output_files}
\title{List postprocessed output files for a stream based on its input spec}
\usage{
get_postproc_output_files(input_dir, input_regex, bids_desc)
}
\arguments{
\item{input_dir}{Directory containing postprocessed outputs.}

\item{input_regex}{Specification used to match the input files for the stream.
May be a space-separated set of BIDS entities or a regex prefixed with "regex:".}

\item{bids_desc}{The \code{desc} value used for postprocessed outputs.}
}
\value{
A character vector of full paths to matching postprocessed outputs.
}
\description{
Converts a postprocess input specification into a pattern that targets
postprocessed outputs by ensuring the \code{desc} entity matches \code{bids_desc}.
}
