% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_functions.R
\name{intensity_normalize}
\alias{intensity_normalize}
\title{Normalize global intensity of a 4D fMRI image}
\usage{
intensity_normalize(
  in_file,
  out_file,
  brain_mask = NULL,
  global_median = 10000,
  overwrite = FALSE,
  lg = NULL,
  fsl_img = NULL
)
}
\arguments{
\item{in_file}{Path to the input 4D NIfTI file.}

\item{out_file}{The full path for the file output by this step}

\item{brain_mask}{Optional path to a brain mask NIfTI file. If \code{NULL}, the entire image is used.}

\item{global_median}{Target median intensity value to normalize to (default is 10000).}

\item{overwrite}{Logical; whether to overwrite the output file if it exists.}

\item{lg}{Optional lgr object used for logging messages}

\item{fsl_img}{Optional Singularity image to execute FSL commands in a containerized environment.}
}
\value{
Path to the intensity-normalized output NIfTI file.
}
\description{
Rescales the intensity of a 4D NIfTI image so that the median voxel intensity within a brain mask
matches a specified global target. This operation is commonly used to standardize signal across runs or subjects.
}
\details{
The 50th percentile intensity is estimated using \code{fslstats}, and the input image is
rescaled using \code{fslmaths -mul}. If the output file exists and \code{overwrite = FALSE}, the step is skipped.
}
\keyword{internal}
