% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_postprocess.R
\name{maybe_add_framewise_displacement}
\alias{maybe_add_framewise_displacement}
\title{Optionally add framewise displacement to confound_calculate columns}
\usage{
maybe_add_framewise_displacement(ppcfg = list(), fields = NULL)
}
\arguments{
\item{ppcfg}{a postprocessing configuration list (nested within scfg$postprocess)}

\item{fields}{Optional vector of fields being edited.}
}
\value{
Modified \code{ppcfg}.
}
\description{
If confound calculation is enabled and framewise displacement was not explicitly
requested, prompt the user to add it. Users can choose whether to use FD
recomputed after motion filtering (when enabled) and whether FD should be
processed with the same filtering/denoising steps as the BOLD data (\code{columns})
or retained as an unprocessed QC covariate (\code{noproc_columns}).
}
\keyword{internal}
