% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle_functions.R
\name{read_project_plan}
\alias{read_project_plan}
\title{Read a saved execution plan}
\usage{
read_project_plan(file)
}
\arguments{
\item{file}{YAML plan path.}
}
\value{
A \code{bg_project_plan} object.
}
\description{
Read a saved execution plan
}
