% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_functions.R
\name{resample_template_to_img}
\alias{resample_template_to_img}
\title{Resample TemplateFlow Mask to fMRIPrep Image Using Python}
\usage{
resample_template_to_img(
  in_file,
  output = NULL,
  template_resolution = 1,
  suffix = "mask",
  desc = "brain",
  extension = ".nii.gz",
  interpolation = "nearest",
  install_dependencies = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{in_file}{Path to the BIDS-compliant NIfTI file.}

\item{output}{Path to output resampled image.}

\item{template_resolution}{Resolution index (e.g., 1 = 1mm).}

\item{suffix}{TemplateFlow suffix (e.g., "mask", "T1w").}

\item{desc}{TemplateFlow descriptor (e.g., "brain").}

\item{extension}{File extension (default: ".nii.gz").}

\item{interpolation}{Interpolation method ("nearest", "linear", "continuous").}

\item{overwrite}{Logical. If \code{TRUE}, overwrite existing \code{output}.}
}
\value{
Invisible \code{TRUE} on success. Writes output to \code{output}.
}
\description{
Resample TemplateFlow Mask to fMRIPrep Image Using Python
}
\details{
The relevant template will be identified using the space- entity for the BIDS-compliant input image.
For example, sub-221256_task-trust_run-1_space-MNI152NLin2009cAsym_desc-preproc_bold.nii.gz
will use the MNI152NLin2009cAsym template image from TemplateFlow.
}
