% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bids_validation.R
\name{run_bids_validation}
\alias{run_bids_validation}
\title{Run BIDS validation on the project directory}
\usage{
run_bids_validation(scfg, outfile = NULL)
}
\arguments{
\item{scfg}{A \code{bg_project_cfg} object containing the project configuration.}

\item{outfile}{Name of the HTML report to create in the BIDS directory. If \code{NULL}, the value from \code{scfg\$bids_validation\$outfile} is used.}}
\value{The job id returned by the scheduler.}
\description{Submit the BIDS validator as a standalone HPC job.}
\examples{
\dontrun{
run_bids_validation(study_config, outfile = "bids_validator_output.html")
}
}
