% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_project.R
\name{run_project}
\alias{run_project}
\title{Run the processing pipeline}
\usage{
run_project(scfg, steps = NULL, prompt = TRUE, debug = FALSE, force = FALSE)
}
\arguments{
\item{scfg}{A list containing the study configuration.}

\item{steps}{Character vector of pipeline steps to execute (or \code{NULL} to run all steps).
Options are c("bids_conversion", "mriqc", "fmriprep", "aroma", "postprocess").}

\item{prompt}{A logical value indicating whether to prompt the user for input on which steps to run.}

\item{debug}{A logical value indicating whether to run in debug mode (verbose output for debugging, no true processing).}

\item{force}{A logical value indicating whether to force the execution of all steps, regardless of their current status.}
}
\value{
A logical value indicating whether the processing pipeline was successfully run.
}
\description{
Run the processing pipeline
}
\examples{
  \dontrun{
    # Assuming you have a valid study configuration list named `study_config`
    run_project(study_config, prompt = TRUE, force = FALSE)
  }
}
