% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_postprocess.R
\name{setup_apply_mask}
\alias{setup_apply_mask}
\title{Configure brain masking for postprocessing}
\usage{
setup_apply_mask(scfg = list(), fields = NULL)
}
\arguments{
\item{scfg}{A study configuration object created by \code{setup_project()}.}

\item{fields}{A character vector of fields to be prompted for. If \code{NULL}, all fields related to brain masking will be prompted.}
}
\value{
A modified version of \code{scfg} with the \verb{$postprocess$apply_mask} entry populated.
}
\description{
This function configures the optional step of applying a brain mask to the functional MRI data
in postprocessing. This step removes signal outside the brain (e.g., in air or non-brain tissue)
by zeroing out voxels outside the specified mask. Users can define a custom mask file or rely on
a default mask derived from the preprocessing pipeline (e.g., fMRIPrep outputs).
}
\details{
This step is especially useful when preparing data for statistical modeling, as it constrains the
analysis to in-brain voxels and reduces computational burden.
}
\keyword{internal}
