% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_project}
\alias{setup_project}
\title{Setup the processing pipeline for a new fMRI study}
\usage{
setup_project(input = NULL, fields = NULL)
}
\arguments{
\item{input}{An existing \code{bg_project_cfg} object to be modified, or a string specifying the location of an existing configuration YAML file to load.}

\item{fields}{A character vector of fields to be prompted for. If \code{NULL}, all fields will be prompted for.}
}
\value{
A \code{bg_project_cfg} list containing the study configuration. New fields are added based on user input,
and missing entries are filled with defaults.
}
\description{
Setup the processing pipeline for a new fMRI study
}
