% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_functions.R
\name{summary.bg_status_df}
\alias{summary.bg_status_df}
\title{Summarize project status}
\usage{
\method{summary}{bg_status_df}(object, ...)
}
\arguments{
\item{object}{A data.frame produced by \code{get_project_status()}.}

\item{...}{Additional arguments (unused)}
}
\value{
data.frame summarizing number of subjects completed for each step.
}
\description{
Provides a tabular summary of completion counts for each step in the pipeline.
}
