% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_project.R
\name{validate_project}
\alias{validate_project}
\title{Validate the structure of a study configuration object}
\usage{
validate_project(scfg = list(), quiet = FALSE)
}
\arguments{
\item{scfg}{a study configuration file as produced by \code{load_project} or \code{setup_project}}
}
\description{
Validate the structure of a study configuration object
}
\keyword{internal}
